Add logout functionality to Header component
Updated the Header.tsx component to include a logout function. This function makes a POST request to the logout endpoint of the API and, if successful, the user is redirected to the home page and the local storage "user" record is removed. The display of the header was also modified to include a logout link when a user is logged in.